Revert "Fix wrong int cast."
This reverts commit cf02b290d702d3d03bafef54a7628b0f0ab22db8.

+++ codecCtx.go
@@ -288,7 +288,7 @@ func (this *CodecCtx) Channels() int {
 }
 
 func (this *CodecCtx) SetBitRate(val int) *CodecCtx {
-	this.avCodecCtx.bit_rate = C.int(val)
+	this.avCodecCtx.bit_rate = C.int64_t(val)
 	return this
 }
